[ADD] Multiple bugs fixed, and new features added
