+++ Assets/Videos/imagen 1.mp4.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4692978c438dee4c9ff711a9bca86f5
+timeCreated: 1513367796
+licenseType: Free
+MovieImporter:
+  serializedVersion: 1
+  quality: 0.5
+  linearTexture: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
