--- Assets/_Memoria/Scripts/Managers/BITalinoManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a2f256cf931523946b3cf5d3157ff49e
-timeCreated: 1511704500
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
