+++ ProjectSettings/AudioManager.asset
@@ -13,3 +13,4 @@ AudioManager:
   m_RealVoiceCount: 32
   m_SpatializerPlugin: 
   m_DisableAudio: 0
+  m_VirtualizeEffects: 1
